+++ engine/src/flutter/flow/testing/mock_embedder.cc
@@ -37,7 +37,7 @@ void MockViewEmbedder::BeginFrame(
 
 // |ExternalViewEmbedder|
 void MockViewEmbedder::PrerollCompositeEmbeddedView(
-    int view_id,
+    int64_t view_id,
     std::unique_ptr<EmbeddedViewParams> params) {
   prerolled_views_.emplace_back(view_id);
 }
@@ -53,7 +53,7 @@ std::vector<DisplayListBuilder*> MockViewEmbedder::GetCurrentBuilders() {
 }
 
 // |ExternalViewEmbedder|
-EmbedderPaintContext MockViewEmbedder::CompositeEmbeddedView(int view_id) {
+EmbedderPaintContext MockViewEmbedder::CompositeEmbeddedView(int64_t view_id) {
   painted_views_.emplace_back(view_id);
   EmbedderPaintContext context = contexts_.front();
   contexts_.pop_front();
